Refactor: routing, pages.
