+++ src/main/java/onlyoffice/com/data/PagesRouting.java
@@ -2,4 +2,5 @@ package onlyoffice.com.data;
 
 public class PagesRouting {
     public static final String LOGIN_PAGE_URL = "https://www.onlyoffice.com/ru/signin.aspx";
+    public static final String MAIN_PAGE_URL = "https://www.onlyoffice.com/";
 }
