+++ src/main/java/onlyoffice/com/site/pages/HomePage.java
@@ -1,4 +1,4 @@
-package onlyoffice.com.pages;
+package onlyoffice.com.site.pages;
 
 public class HomePage {
 }
