+++ src/main/java/onlyoffice/com/site/pages/MainPage.java
@@ -1,6 +1,8 @@
-package onlyoffice.com.pages;
+package onlyoffice.com.site.pages;
 
-import onlyoffice.com.pages.components.homePage.HeaderBlock;
+import io.qameta.allure.Step;
+import onlyoffice.com.data.PagesRouting;
+import onlyoffice.com.site.components.HeaderBlock;
 
 import static com.codeborne.selenide.Selenide.$x;
 import static com.codeborne.selenide.Selenide.open;
@@ -10,13 +12,13 @@ public class MainPage {
 
     private HeaderBlock headerBlock;
 
-    public static final String MAIN_PAGE_URL = "https://www.onlyoffice.com/";
-
+    @Step
     public static MainPage enterToMainPage(){
-        open(MAIN_PAGE_URL);
+        open(PagesRouting.MAIN_PAGE_URL);
         return page(MainPage.class);
     }
 
+    @Step
     public LoginPage clickOnLoginButton(){
         $x(headerBlock.LOGIN_BTN_XPATH).click();
         return page(LoginPage.class);
