+++ src/test/java/onlyoffice/com/tests/LoginTest.java
@@ -1,13 +1,12 @@
 package onlyoffice.com.tests;
 
-import onlyoffice.com.steps.LoginStep;
+import onlyoffice.com.steps.NavigationSteps;
 import org.testng.annotations.Test;
 
 public class LoginTest extends BaseTest {
 
     @Test
     public void loginTest(){
-        LoginStep loginStep = new LoginStep();
-        loginStep.login("enter@your.email", "password");
+        NavigationSteps.loginWithValidCredentials();
     }
 }
